Conflicts:
	WebContent/WEB-INF/jsp/base/bill/recheckNotes.jsp
	WebContent/WEB-INF/jsp/base/spectype/addspectype.jsp
	WebContent/WEB-INF/jsp/sysLogin.jsp
	WebContent/dwz/js/dwz.ajax.js
	WebContent/dwz/js/dwz.core.js
	WebContent/dwz/js/dwz.stable.js
	WebContent/dwz/themes/purple/style.css
	WebContent/dwz/themes/silver/style.css
	WebContent/oper/js/om-itemselector.js
	WebContent/oper/js/om-sortable.js
	WebContent/oper/om-default.css
	src/com/ylbms/base/bill/web/controller/BillController.java
	src/com/ylbms/base/check/dao/CheckNotesInfoDao.java
	src/com/ylbms/base/check/model/CheckNotesInfo.java
	src/com/ylbms/base/check/model/CheckNotesInfoPK.java
	src/com/ylbms/base/check/web/controller/CheckNotesController.java
	src/com/ylbms/base/check/web/controller/NotesDTO.java
	src/com/ylbms/base/location/util/DBContextHolder.java
	src/com/ylbms/base/location/util/DynamicDataSource.java
	src/com/ylbms/base/single/dao/SingleInfoDao.java
	src/com/ylbms/base/single/model/SingleDTO.java
	src/com/ylbms/base/single/service/SpectypeService.java
	src/com/ylbms/common/interceptor/SessionInterceptor.java
	src/com/ylbms/common/utils/Collections3.java
	src/com/ylbms/common/utils/SpringContextHolder.java
	src/com/ylbms/common/utils/StringUtils.java
	src/com/ylbms/common/web/OrgTreeTag.java
	src/com/ylbms/system/security/SystemRealm.java
	src/com/ylbms/system/service/impl/SystemServiceImpl.java
	src/com/ylbms/system/utils/UserUtils.java
	src/com/ylbms/system/web/controller/DictController.java
	src/com/ylbms/system/web/controller/RoleController.java
	src/com/ylbms/test/UserServiceTest.java

+++ src/com/ylbms/system/web/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.ylbms.system.web.controller;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -10,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.apache.shiro.authz.annotation.RequiresUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -20,6 +20,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.google.common.collect.Lists;
 import com.ylbms.common.orm.Page;
 import com.ylbms.common.orm.PropertyFilter;
 import com.ylbms.common.utils.DwzUtil;
@@ -28,6 +29,7 @@ import com.ylbms.system.model.Role;
 import com.ylbms.system.model.User;
 import com.ylbms.system.service.SystemService;
 import com.ylbms.system.service.UserSerivice;
+import com.ylbms.system.utils.UserUtils;
 
 /**
  * 
@@ -64,7 +66,7 @@ public class UserController extends BaseController {
 	 * @param model
 	 * @return
 	 */
-	@RequiresUser
+	@RequiresPermissions("sys:user:add")
 	@RequestMapping("/addUi")
 	public String addUi(HttpServletRequest request, Model model) {
 		model.addAttribute("allRoles", systemService.findAllRole());
@@ -114,7 +116,7 @@ public class UserController extends BaseController {
 	 * @param model
 	 * @return
 	 */
-	@RequiresUser
+	@RequiresPermissions("sys:user:edit")
 	@RequestMapping(value = "/editUi/{id}")
 	public String editUi(HttpServletRequest request,
 			@PathVariable("id") Long id, Model model) {
@@ -135,13 +137,13 @@ public class UserController extends BaseController {
 	public Map<String, Object> addRole(HttpServletRequest request, User user,
 			String roleIds) {
 		try {
-			List<Role> roleList = new ArrayList<Role>();
+			List<Role> roleList =Lists.newArrayList();
 			String[] role = roleIds.split(",");
 			for (int i = 0, len = role.length; i < len; i++) {
 				roleList.add(new Role(Long.parseLong(role[i])));
 			}
 			user.setRoleList(roleList);
-			systemService.saveUser(user);
+			systemService.updateUser(user);
 		} catch (Exception e) {
 			log.error("system error!!", e);
 			return DwzUtil.dialogAjaxDone(DwzUtil.FAIL, "user", e.getMessage());
@@ -176,9 +178,10 @@ public class UserController extends BaseController {
 	@ResponseBody
 	public Map<String, Object> updateUser(User user, String oldPwd) {
 		try {
-			if(StringUtils.isNotBlank(user.getPassword())){
-				user.setPassword(systemService.entryptPassword(user.getPassword()));
-			}else{
+			if (StringUtils.isNotBlank(user.getPassword())) {
+				user.setPassword(systemService.entryptPassword(user
+						.getPassword()));
+			} else {
 				user.setPassword(oldPwd);
 			}
 			systemService.updateUser(user);
@@ -195,17 +198,25 @@ public class UserController extends BaseController {
 	 * @param id
 	 * @return
 	 */
-	@RequiresUser
+	@RequiresPermissions("sys:user:delete")
 	@RequestMapping(value = "/delete/{id}")
 	@ResponseBody
 	public Map<String, Object> delete(@PathVariable("id") Long id) {
 		try {
+			if (UserUtils.getUser().getId().equals(id)) {
+				return DwzUtil.dialogAjaxDone(DwzUtil.FAIL, null,
+						"删除用户失败，不能删除当前用户！");
+			}
+			if (User.isAdmin(id)) {
+				return DwzUtil.dialogAjaxDone(DwzUtil.FAIL, null,
+						"删除用户失败，不允许删除超级管理员");
+			}
 			userService.deleteUser(id);
+			return DwzUtil.dialogAjaxDone(DwzUtil.OK);
 		} catch (Exception e) {
 			log.error("system error!!", e);
 			return DwzUtil.dialogAjaxDone(DwzUtil.FAIL, "user", e.getMessage());
 		}
-		return DwzUtil.dialogAjaxDone(DwzUtil.OK);
 	}
 
 	/**
@@ -214,7 +225,7 @@ public class UserController extends BaseController {
 	 * @param ids
 	 * @return
 	 */
-	@RequiresUser
+	@RequiresPermissions("sys:user:delete")
 	@RequestMapping(value = "/delByIds/{ids}")
 	@ResponseBody
 	public Map<String, Object> delByIds(@RequestParam("ids") String ids) {
